Rename to 'customunits' instead of 'uagunits'
