+++ mods/uagunits/addons/customunits/CfgFactionClasses.hpp
@@ -1,0 +1,6 @@
+class CfgFactionClasses {
+    class BLU_F;
+    class UAGPMC: BLU_F {
+        displayName = "UAGPMC";
+    };
+};
