+++ mods/uagunits/addons/customunits/CfgGlasses.hpp
@@ -1,0 +1,8 @@
+class CfgGlasses {
+	class G_Balaclava_TI_G_blk_F;
+
+	class UAGPMC_Balaclava_Goggles: G_Balaclava_TI_G_blk_F {
+		displayName = "UAG Balaclava w/ Goggles";
+		hiddenSelectionsTextures[] = {"\A3\Characters_F_Exp\BLUFOR\Data\G_Balaclava_TI_blk_F_co.paa","\A3\Characters_F_Exp\BLUFOR\Data\G_Balaclava_TI_blk_F_co.paa"};
+	};
+};
