+++ mods/uagunits/addons/customunits/CfgGroups.hpp
@@ -1,0 +1,219 @@
+class CfgGroups {
+    class West {
+        class UAGPMC {
+            name = "UAGPMC";
+
+            class Wolfpack {
+                name = "Wolfpack";
+
+                class UAGPMC_Wolfpack_Actual {
+                    name = "Actual (Command)";
+                    faction = "UAGPMC";
+                    icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+
+                    class Unit0 {
+                        position[] = {0,0,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_SectionCommander";
+                    };
+
+                    class Unit2 {
+                        position[] = {-5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Medic";
+                    };
+                };
+
+                class UAGPMC_Wolfpack_Charlie {
+                    name = "Charlie (Base-of-Fire)";
+                    faction = "UAGPMC";
+                    icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+
+                    class Unit0 {
+                        position[] = {0,0,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Teamleader";
+                    };
+
+                    class Unit1 {
+                        position[] = {5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Corpsman";
+                    };
+
+                    class Unit2 {
+                        position[] = {-5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Marksman";
+                    };
+
+                    class Unit3 {
+                        position[] = {10,-10,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Autorifleman";
+                    };
+
+                    class Unit4 {
+                        position[] = {-10,-10,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_BitchBoy";
+                    };
+                };
+
+                class UAGPMC_Wolfpack_Delta {
+                    name = "Delta (Assault)";
+                    faction = "UAGPMC";
+                    icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+
+                    class Unit0 {
+                        position[] = {0,0,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Teamleader";
+                    };
+
+                    class Unit1 {
+                        position[] = {5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Corpsman";
+                    };
+
+                    class Unit2 {
+                        position[] = {-5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Pointman";
+                    };
+
+                    class Unit3 {
+                        position[] = {10,-10,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Raider";
+                    };
+
+                    class Unit4 {
+                        position[] = {-10,-10,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Contractor";
+                    };
+                };
+            };
+
+            class Sabre {
+                name = "Sabre";
+
+                class UAGPMC_Sabre_Actual {
+                    name = "Actual (Command)";
+                    faction = "UAGPMC";
+                    icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+
+                    class Unit0 {
+                        position[] = {0,0,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_SectionCommander";
+                    };
+
+                    class Unit2 {
+                        position[] = {-5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Medic";
+                    };
+                };
+
+                class UAGPMC_Sabre_Charlie {
+                    name = "Charlie (Weapons)";
+                    faction = "UAGPMC";
+                    icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+
+                    class Unit0 {
+                        position[] = {0,0,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Teamleader";
+                    };
+
+                    class Unit1 {
+                        position[] = {5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Corpsman";
+                    };
+
+                    class Unit2 {
+                        position[] = {-5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Sniper";
+                    };
+
+                    class Unit3 {
+                        position[] = {10,-10,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Machinegunner";
+                    };
+
+                    class Unit4 {
+                        position[] = {-10,-10,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_BitchBoy";
+                    };
+                };
+
+                class UAGPMC_Sabre_Delta {
+                    name = "Delta (Assault)";
+                    faction = "UAGPMC";
+                    icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+
+                    class Unit0 {
+                        position[] = {0,0,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Teamleader";
+                    };
+
+                    class Unit1 {
+                        position[] = {5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Corpsman";
+                    };
+
+                    class Unit2 {
+                        position[] = {-5,-5,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Pointman";
+                    };
+
+                    class Unit3 {
+                        position[] = {10,-10,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Raider";
+                    };
+
+                    class Unit4 {
+                        position[] = {-10,-10,0};
+                        rank = "PRIVATE";
+                        side = 1;
+                        vehicle = "UAGPMC_Grenadier";
+                    };
+                };
+            };
+        };
+    };
+};
