+++ mods/uagunits/addons/customunits/CfgPatches.hpp
@@ -1,0 +1,24 @@
+class CfgPatches {
+    class uagpmc_customunits {
+        name = "uagpmc_customunits";
+        units[] = {
+            "UAGPMC_Survivor",
+            "UAGPMC_Contractor",
+            "UAGPMC_Autorifleman",
+            "UAGPMC_Machinegunner",
+            "UAGPMC_Pointman",
+            "UAGPMC_Corpsman",
+            "UAGPMC_Medic",
+            "UAGPMC_Teamleader",
+            "UAGPMC_SectionCommander",
+        };
+        requiredAddons[] = {
+            "A3_weapons_f_exp",
+            "Splinter_camo_addon",
+            "cTab",
+            "ace_common"
+        };
+        authors[] = {"uagpmc.com", "zuedev"};
+        authorUrl = "https://uagpmc.com";
+    };
+};
