+++ mods/uagunits/addons/customunits/config.cpp
@@ -1,0 +1,7 @@
+#include "macros.hpp"
+#include "CfgPatches.hpp"
+#include "CfgWeapons.hpp"
+#include "CfgGlasses.hpp"
+#include "CfgFactionClasses.hpp"
+#include "CfgVehicles.hpp"
+#include "CfgGroups.hpp"
