--- mods/uagunits/addons/uagunits/$PBOPREFIX$
@@ -1,1 +1,0 @@
-uagpmc_uagunits
