--- mods/uagunits/addons/uagunits/CfgFactionClasses.hpp
@@ -1,6 +1,0 @@
-class CfgFactionClasses {
-    class BLU_F;
-    class UAGPMC: BLU_F {
-        displayName = "UAGPMC";
-    };
-};
