--- mods/uagunits/addons/uagunits/CfgPatches.hpp
@@ -1,24 +1,0 @@
-class CfgPatches {
-    class uagpmc_uagunits {
-        name = "uagunits";
-        units[] = {
-            "UAGPMC_Survivor",
-            "UAGPMC_Contractor",
-            "UAGPMC_Autorifleman",
-            "UAGPMC_Machinegunner",
-            "UAGPMC_Pointman",
-            "UAGPMC_Corpsman",
-            "UAGPMC_Medic",
-            "UAGPMC_Teamleader",
-            "UAGPMC_SectionCommander",
-        };
-        requiredAddons[] = {
-            "A3_weapons_f_exp",
-            "Splinter_camo_addon",
-            "cTab",
-            "ace_common"
-        };
-        authors[] = {"uagpmc.com", "zuedev"};
-        authorUrl = "https://uagpmc.com";
-    };
-};
