--- mods/uagunits/addons/uagunits/macros.hpp
@@ -1,10 +1,0 @@
-#define x2(x) x, x
-#define x3(x) x, x, x
-#define x4(x) x, x, x, x
-#define x5(x) x, x, x, x, x
-#define x6(x) x, x, x, x, x, x
-#define x7(x) x, x, x, x, x, x, x
-#define x8(x) x, x, x, x, x, x, x, x
-#define x9(x) x, x, x, x, x, x, x, x, x
-#define x10(x) x, x, x, x, x, x, x, x, x, x
-#define x20(x) x, x, x, x, x, x, x, x, x, x, x, x, x, x, x, x, x, x, x, x
